requirements.txt
# Core dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0

# Email & HTML parsing
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Document parsing
pypdf>=4.0.0
langchain-community>=0.3.0

# LLM providers
openai>=1.0.0

# Messaging
twilio>=9.0.0

# Configuration
python-dotenv>=1.0.0

# Type hints
typing-extensions>=4.9.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development tools
black>=24.0.0
isort>=5.13.0
flake8>=7.0.0
